+++ crates/rattler_solve/src/resolvo/mod.rs
@@ -165,7 +165,11 @@ impl SolverPackageRecord<'_> {
 
     fn timestamp(&self) -> Option<&chrono::DateTime<chrono::Utc>> {
         match self {
-            SolverPackageRecord::Record(rec) => rec.package_record.timestamp.as_ref(),
+            SolverPackageRecord::Record(rec) => rec
+                .package_record
+                .timestamp
+                .as_ref()
+                .map(TimestampMs::datetime),
             SolverPackageRecord::Extra { .. } | SolverPackageRecord::VirtualPackage(..) => None,
         }
     }